+++ _scripts/set-version.sh
@@ -1,22 +1,30 @@
 version="0.11.0"
 
-for i in tranalyze trcombine trconvert trdelabel trdelete trdot trfold trfoldlit trformat trgen trgroup trinsert trjson trkleene trmvsr trparse trprint trrename trrr trrup trsplit trsponge trst trstrip trtext trthompson trtokens trtree trunfold trungroup trwdog trxgrep trxml trxml2
+directories=`find . -maxdepth 1 -type d`
+for i in $directories
 do
+	if [ "$i" == "." ]
+	then
+		continue
+	fi
+	cd $i
+	csproj=`find . -maxdepth 1 -name '*.csproj'`
+	if [[ "$csproj" == "" ]]
+	then
+		cd ..
+		continue
+	fi
+	if [[ ! -f "$i.csproj" ]]
+	then
+		echo $i
+		echo nope
+		exit 1
+	fi
 	echo $i
-	pushd $i
 	rm -f asdfasdf
 	cat *.csproj | sed -e "s%[<][Vv]ersion[>].*[<][/][Vv]ersion[>]%<Version\>$version</Version>%" > asdfasdf
 	mv asdfasdf *.csproj	
-	popd
-done
-for i in tragl
-do
-	echo $i
-	pushd $i
-	rm -f asdfasdf
-	cat *.csproj | sed -e "s%[<][Vv]ersion[>].*[<][/][Vv]ersion[>]%<Version\>$version</Version>%" > asdfasdf
-	mv asdfasdf *.csproj	
-	popd
+	cd ..
 done
 
 pushd trgen
